Rapid7InsightVM/pyproject.toml
[tool.poetry]
name = "Automation module for Rapid7insightvm"
version = "1.0.0"
description = "Rapid7 InsightVM connector for Sekoia"
authors = []
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
sekoia-automation-sdk = "^1.22.5"
requests = "^2.31"

[tool.poetry.dev-dependencies]
pytest = "*"
pytest-cov = "*"
requests-mock = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 119
target-version = ['py38']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 119

[tool.coverage.run]
omit = [
    "tests/*",
    "main.py",
]

[tool.pytest.ini_options]
minversion = "6.0"
testpaths = [
    "tests",
]
addopts = '''
  --strict-markers
  --tb=short
  --cov=rapid7insightvm_modules
  --cov-branch
  --cache-clear
  --cov-report=html
  --cov-report=term-missing:skip-covered
  --cov-fail-under=5
  --capture=sys
'''
